pyproject.toml
[tool.poetry]
name = "home-work-web-13"
version = "0.1.0"
description = "my home work"
authors = ["andriymyr <andriymyr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
alembic = "^1.13.0"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
uvicorn = "^0.24.0.post1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
pydantic = { extras = ["email"], version = "^2.5.2" }
python-multipart = "^0.0.6"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.0"
redis = ">=4.0.0,<5.0.0"
fastapi-limiter = "^0.1.5"
jinja2 = "^3.1.2"
cloudinary = "^1.37.0"
psycopg2 = "^2.9.9"
aioredis = "^2.0.1"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
httpx = "^0.26.0"
aiosqlite = "^0.19.0"



[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
